+++ src/main/java/lotto/constants/LottoConstants.java
@@ -1,6 +1,6 @@
 package lotto.constants;
 
-public class Constants {
+public class LottoConstants {
     public static final int LOTTO_PRICE = 1000;
     public static final int LOTTO_NUMBER_COUNT = 6;
     public static final int MIN_LOTTO_NUMBER = 1;
